doom: make MEETING not a todo item
It did not make much sense anyway. It was considered as a "DONE"
todo item, even it was in the future. Also the "MEETING" string
was intepreted as a todo tag and then removed from the title for
calendar sync, so "MEETING with Person X" had a corresponding
calendar entry "with Person X". This change fixes that behavior.