doom: projectile fixes
Previously, I used the alien indexing method which used .gitignore to
ignore files from projectile-find-file. I was very confused why some of
my files would not show up in the search. Switching to the native
indexing method seems to fix that, as well as the sort-order which is
ignored by alien, because alien uses external tools to list files and
these tools do not know abotu that emacs variable.

Actually it would make much more sense if projectile would use the
.projectile file when it is present, and only if it is not use
.gitignore. Unfortunately, this is an open issue:
https://stackoverflow.com/questions/63432901/projectile-use-projectile-instead-of-gitignore-for-excluding-file-from-proje